fix: Fix npm start error
