+++ src/App.js
@@ -1,19 +1,18 @@
-import React, { Component } from 'react';
-import './App.css';
+import React from "react";
+import "./App.css";
 // import Button from './button';
-import Input from './input/input';
-import Choicelist from './choices';
-  
+import Input from "./input/input";
+// import Choicelist from './choices';
+
 class App extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      enterInput: 'Input'
-    }
+      enterInput: "Input",
+    };
 
     // handleChange = (event) => {
     //   const input = event.target.name;
-
 
     //   this.setState({
     //     enterInput: event.target.value
@@ -21,17 +20,13 @@ class App extends React.Component {
     // }
   }
 
-
   render() {
-    return (<div>
-      <Input 
-      name="enterInput"
-      
-      />
-    </div>
-      
-      )
-  };
+    return (
+      <div>
+        <Input name="enterInput" />
+      </div>
+    );
+  }
 }
 
 export default App;
