+++ src/input/input.js
@@ -1,18 +1,18 @@
 import React from "react";
 // import LinkButton from "../linkButton/linkbutton";
-import Button from '../button/button';
-import '../input/input.css';
-import ChoiceList from '../choices';
+import Button from "../button/button";
+import "../input/input.css";
+import ChoiceList from "../choices";
 
 class Input extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      question: '',
-      firstOption: '',
-      secondOption: '',
-      thirdOption: '',
-      fourthOption: ''
+      question: "",
+      firstOption: "",
+      secondOption: "",
+      thirdOption: "",
+      fourthOption: "",
     };
   }
 
@@ -23,22 +23,24 @@ class Input extends React.Component {
     //   event.target.secondOption,
     //   event.target.thirdOption,
     //   event.target.fourthOption, ];
-    const  options = [ option1, option2, option3, option4] = [
+
+    let option1 = "";
+    let option2 = "";
+    let option3 = "";
+    let option4 = "";
+
+    const options = [option1, option2, option3, option4] || [
       event.target.firstOption,
       event.target.secondOption,
       event.target.thirdOption,
-      event.target.fourthOption, ];
-      let option1 = "";
-      let option2= "";
-      let option3 = "";
-      let option4= ""
+      event.target.fourthOption,
+    ];
 
     let val = event.target.value;
 
     this.setState({
       [quest]: val,
-      options
-
+      options,
 
       // [option1]: val,
       // [option2]: val,
@@ -46,8 +48,6 @@ class Input extends React.Component {
       // [option4]: val
     });
   };
-
-  
 
   render() {
     return (
@@ -57,7 +57,7 @@ class Input extends React.Component {
           <p>Enter your question and options in the input placeholder below</p>
         </header>
 
-        <section className="section" role="">
+        <section className="section">
           <form action="href" className="choice-container">
             <input
               onChange={this.myChangeHandler}
@@ -102,19 +102,14 @@ class Input extends React.Component {
               }}
           ></LinkButton> */}
         </section>
-        <Button
-          onClick={this.myChangeHandler.bind(this)}
-          
-        />
+        <Button onClick={this.myChangeHandler.bind(this)} />
         <ChoiceList
-          quest={this.state.question} 
+          quest={this.state.question}
           option1={this.state.firstOption}
           option2={this.state.secondOption}
           option3={this.state.thirdOption}
           option4={this.state.fourthOption}
-        
         ></ChoiceList>
-
       </div>
     );
   }
